Search: merge snapshot integration test into bleve_snapshot_test.go
Grafana's convention is to identify integration tests by the
TestIntegration prefix, not by a separate _integration_test.go file.
Renamed TestBleveBackend_SnapshotRoundTrip to
TestIntegrationBleveSnapshotRoundTrip and folded it + its helpers
into the unit test file.
